Fixed unit_price issue, refactored rake import file, changed u_p to decimal with calculate_money private method in items and invoice_items model

+++ db/migrate/20160223000441_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :unit_price
+      t.decimal :unit_price
       t.references :merchant, index: true, foreign_key: true
 
       t.timestamps null: false
